feat: Контекстное меню + разбил MapCanvas
Разбил MapCanvas на хук, файлы управление камерой и отрисовки зданий. контекстное меню, которое прикрепленно к точке нажатия
